refactor: split LLM analysis API into smaller feature-specific functions and temporary deactivation of the streaming function
