+++ python/insight/app/api/llm_analysis/__init__.py
@@ -1,0 +1,13 @@
+from .controller.llm_session import router as session_router
+from .controller.api_key import router as api_key_router
+from .controller.llm_model import router as model_router
+from .controller.log_analysis import router as log_analysis_router
+from .controller.alarm_analysis import router as alarm_analysis_router
+
+__all__ = [
+    "session_router",
+    "api_key_router",
+    "model_router",
+    "log_analysis_router",
+    "alarm_analysis_router"
+]
