--- python/insight/app/api/log_analysis/utils/utils.py
@@ -1,281 +1,0 @@
-from app.api.log_analysis.response.res import (
-    LogAnalysisModel,
-    LogAnalysisSession,
-    SessionHistory,
-    Message,
-    OpenAIAPIKey,
-    GoogleAPIKey,
-    QueryMetadata,
-    AnthropicAPIKey,
-)
-from app.api.log_analysis.repo.repo import LogAnalysisRepository
-from app.api.log_analysis.request.req import PostSessionBody, SessionIdPath, PostQueryBody
-from app.core.mcp.mcp_context import MCPContext
-from app.core.mcp.multi_mcp_manager import MCPManager
-from sqlalchemy.orm import Session
-from fastapi import HTTPException, status
-from typing import Callable
-import os
-import uuid
-
-
-class LogAnalysisService:
-    PROVIDER_ENV_MAP = {
-        "ollama": "OLLAMA_BASE_URL",
-        "openai": "OPENAI_API_KEY",
-        "google": "GOOGLE_API_KEY",
-        "anthropic": "ANTHROPIC_API_KEY",
-    }
-
-    def __init__(self, db: Session = None, mcp_context=None):
-        self.repo = LogAnalysisRepository(db=db)
-        # mcp_context now receives MCPManager instance
-        if isinstance(mcp_context, MCPManager):
-            # Wrap MCPManager with MCPContext
-            self.mcp_context = MCPContext(mcp_context)
-        else:
-            self.mcp_context = mcp_context
-
-    def get_model_list(self, model_info_config):
-        result = []
-        for model_info in model_info_config:
-            env_key = self.PROVIDER_ENV_MAP[model_info["provider"]]
-            if model_info["provider"] == "ollama" and os.getenv(env_key):
-                result.append(self.map_model_to_res(model_info))
-            elif model_info["provider"] == "openai" and self.repo.get_openai_key():
-                result.append(self.map_model_to_res(model_info))
-            elif model_info["provider"] == "google" and self.repo.get_google_key():
-                result.append(self.map_model_to_res(model_info))
-            elif model_info["provider"] == "anthropic" and self.repo.get_anthropic_key():
-                result.append(self.map_model_to_res(model_info))
-            else:
-                pass
-        return result
-
-    @staticmethod
-    def map_model_to_res(model_info):
-        return LogAnalysisModel(provider=model_info["provider"], model_name=model_info["model_name"])
-
-    def get_sessions(self):
-        sessions = self.repo.get_all_sessions()
-        results = [self.map_session_to_res(session) for session in sessions]
-        return results
-
-    def create_chat_session(self, body: PostSessionBody):
-        provider, model_name = body.provider, body.model_name
-        session_id = uuid.uuid4()
-
-        session_info = {"USER_ID": 1, "SESSION_ID": session_id, "PROVIDER": provider, "MODEL_NAME": model_name}
-        new_session = self.repo.create_session(session_info)
-
-        return self.map_session_to_res(new_session)
-
-    @staticmethod
-    def map_session_to_res(session):
-        return LogAnalysisSession(
-            seq=session.SEQ, user_id=session.USER_ID, session_id=session.SESSION_ID, provider=session.PROVIDER, model_name=session.MODEL_NAME, regdate=session.REGDATE
-        )
-
-    def delete_chat_session(self, path: SessionIdPath):
-        session_id = path.sessionId
-        session = self.repo.delete_session_by_id(session_id)
-
-        if session:
-            deleted_session = self.map_session_to_res(session)
-            return deleted_session
-        else:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Session Not Found")
-
-    def delete_all_chat_sessions(self):
-        sessions = self.repo.delete_all_sessions()
-        if sessions:
-            return [self.map_session_to_res(session) for session in sessions]
-        else:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No Sessions Found")
-
-    async def get_chat_session_history(self, path: SessionIdPath):
-        session_id = path.sessionId
-        session_info = self.repo.get_session_by_id(session_id)
-
-        history = await self.mcp_context.get_chat_history(session_id)
-        result = []
-
-        if not session_info:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Session Not Found")
-
-        if history:
-            channel_values = history.get("channel_values", [])
-            for message in channel_values["messages"]:
-                if self.filter_message(message):
-                    result.append(Message(message_type=message.type, message=message.content))
-        return self.map_history_to_res(session_info, result)
-
-    @staticmethod
-    def filter_message(element):
-        return element.type == "human" or (element.type == "ai" and element.content)
-
-    @staticmethod
-    def map_history_to_res(session, messages):
-        return SessionHistory(
-            seq=session.SEQ,
-            user_id=session.USER_ID,
-            session_id=session.SESSION_ID,
-            provider=session.PROVIDER,
-            model_name=session.MODEL_NAME,
-            regdate=session.REGDATE,
-            messages=messages,
-        )
-
-    async def query(self, body: PostQueryBody):
-        session_id, message = body.session_id, body.message
-        session = self.repo.get_session_by_id(session_id)
-        if not session:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Session Not Found")
-        provider_credential = CredentialService(repo=self.repo).get_provider_credential(provider=session.PROVIDER)
-        await self.mcp_context.get_agent(session.PROVIDER, session.MODEL_NAME, provider_credential, streaming=False)
-
-        query_result = await self.mcp_context.aquery(session_id, message)
-        result = query_result["messages"][-1].content
-
-        # Get metadata summary (None if not available)
-        metadata_summary = None
-        try:
-            metadata_summary = self.mcp_context.get_metadata_summary()
-        except Exception:
-            metadata_summary = None
-
-        if not metadata_summary or (
-            not metadata_summary.get("queries_executed")
-            and not metadata_summary.get("total_execution_time")
-            and not metadata_summary.get("tool_calls_count")
-            and not metadata_summary.get("databases_accessed")
-        ):
-            # Forced metadata for visibility confirmation (temporary)
-            metadata_model = QueryMetadata(
-                queries_executed=["SHOW DATABASES"],
-                total_execution_time=0.85,
-                tool_calls_count=1,
-                databases_accessed=["InfluxDB"],
-            )
-        else:
-            metadata_model = QueryMetadata(**metadata_summary)
-
-        return Message(message_type="ai", message=result, metadata=metadata_model)
-
-    async def query_stream(self, body: PostQueryBody):
-        """Prepare agent and return async generator for SSE streaming."""
-        session_id, message = body.session_id, body.message
-        session = self.repo.get_session_by_id(session_id)
-        if not session:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Session Not Found")
-
-        provider_credential = CredentialService(repo=self.repo).get_provider_credential(provider=session.PROVIDER)
-        await self.mcp_context.get_agent(session.PROVIDER, session.MODEL_NAME, provider_credential, streaming=True)
-
-        # Return async generator for StreamingResponse
-        return self.mcp_context.astream_query(session_id, message)
-
-
-class CredentialService:
-    def __init__(self, repo):
-        self.repo = repo
-        self._fetchers: dict[str, Callable[[], str]] = {
-            "openai": self._fetch_openai_key,
-            "ollama": self._fetch_ollama_url,
-            "google": self._fetch_google_key,
-            "anthropic": self._fetch_anthropic_key,
-        }
-
-    def get_provider_credential(self, provider: str) -> str:
-        return self._fetchers[provider]()
-
-    def _fetch_openai_key(self) -> str:
-        api_key = self.repo.get_openai_key()
-        if not api_key:
-            raise HTTPException(detail="API Key Not Found", status_code=status.HTTP_404_NOT_FOUND)
-        return api_key.API_KEY
-
-    @staticmethod
-    def _fetch_ollama_url() -> str:
-        url = os.getenv("OLLAMA_BASE_URL")
-        if not url:
-            raise HTTPException(detail="OLLAMA_BASE_URL not set", status_code=status.HTTP_500_INTERNAL_SERVER_ERROR)
-        return url
-
-    def _fetch_google_key(self) -> str:
-        api_key = self.repo.get_google_key()
-        if not api_key:
-            raise HTTPException(detail="API Key Not Found", status_code=status.HTTP_404_NOT_FOUND)
-        return api_key.API_KEY
-
-    def _fetch_anthropic_key(self) -> str:
-        api_key = self.repo.get_anthropic_key()
-        if not api_key:
-            raise HTTPException(detail="API Key Not Found", status_code=status.HTTP_404_NOT_FOUND)
-        return api_key.API_KEY
-
-
-class OpenAIAPIKeyService:
-    def __init__(self, db: Session = None):
-        self.repo = LogAnalysisRepository(db=db)
-
-    def get_key(self):
-        result = self.repo.get_openai_key()
-        if not result:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No API key found")
-        return OpenAIAPIKey(seq=result.SEQ, api_key=result.API_KEY)
-
-    def post_key(self, api_key: str):
-        result = self.repo.create_openai_key(api_key)
-        return OpenAIAPIKey(seq=result.SEQ, api_key=result.API_KEY)
-
-    def delete_key(self):
-        result = self.repo.get_openai_key()
-        if not result:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No API key to delete")
-        self.repo.delete_openai_key()
-        return OpenAIAPIKey(seq=result.SEQ, api_key=result.API_KEY)
-
-
-class GoogleAPIKeyService:
-    def __init__(self, db: Session = None):
-        self.repo = LogAnalysisRepository(db=db)
-
-    def get_key(self):
-        result = self.repo.get_google_key()
-        if not result:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No API key found")
-        return GoogleAPIKey(seq=result.SEQ, api_key=result.API_KEY)
-
-    def post_key(self, api_key: str):
-        result = self.repo.create_google_key(api_key)
-        return GoogleAPIKey(seq=result.SEQ, api_key=result.API_KEY)
-
-    def delete_key(self):
-        result = self.repo.get_google_key()
-        if not result:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No API key to delete")
-        self.repo.delete_google_key()
-        return GoogleAPIKey(seq=result.SEQ, api_key=result.API_KEY)
-
-
-class AnthropicAPIKeyService:
-    def __init__(self, db: Session = None):
-        self.repo = LogAnalysisRepository(db=db)
-
-    def get_key(self):
-        result = self.repo.get_anthropic_key()
-        if not result:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No API key found")
-        return AnthropicAPIKey(seq=result.SEQ, api_key=result.API_KEY)
-
-    def post_key(self, api_key: str):
-        result = self.repo.create_anthropic_key(api_key)
-        return AnthropicAPIKey(seq=result.SEQ, api_key=result.API_KEY)
-
-    def delete_key(self):
-        result = self.repo.get_anthropic_key()
-        if not result:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="No API key to delete")
-        self.repo.delete_anthropic_key()
-        return AnthropicAPIKey(seq=result.SEQ, api_key=result.API_KEY)
