+++ python/insight/config/ConfigManager.py
@@ -42,7 +42,8 @@ class ConfigManager:
 
     def get_db_config(self):
         db = self.config.get("common", {}).get("DB", {})
-        return {"url": db.get("URL", "localhost"), "user": db.get("USERNAME", "mcmp"), "pw": db.get("PASSWORD", "1234"), "db": db.get("DATABASE", "mcmp")}
+        return {"url": db.get("URL", "localhost"), "user": db.get("USERNAME", "mcmp"), "pw": db.get("PASSWORD", "1234"),
+                "db": db.get("DATABASE", "mcmp")}
 
     def get_influxdb_config(self):
         influxdb = self.config.get("common", {}).get("InfluxDB", {})
@@ -69,18 +70,30 @@ class ConfigManager:
         o11y = self.config.get("common", {}).get("MC-O11Y", {})
         return {"url": o11y.get("URL", ""), "port": o11y.get("PORT", "")}
 
-    def get_model_config(self):
-        model = self.config.get("log_analysis", {}).get("model", [])
-
+    def get_llm_model_config(self):
+        model = self.config.get("llm", {}).get("model", [])
         return model
 
     def get_mcp_config(self):
-        mcp = self.config.get("log_analysis", {}).get("mcp", {})
-        return {"mcp_grafana_url": mcp.get("mcp_grafana_url", ""), "mcp_mariadb_url": mcp.get("mcp_mariadb_url", ""), "mcp_influxdb_url": mcp.get("mcp_influxdb_url", "")}
+        mcp = self.config.get("llm", {}).get("mcp", {})
+        return {"mcp_grafana_url": mcp.get("mcp_grafana_url", ""),
+                "mcp_mariadb_url": mcp.get("mcp_mariadb_url", ""),
+                "mcp_influxdb_url": mcp.get("mcp_influxdb_url", "")}
 
-    def get_system_prompt_config(self):
+    def get_log_system_prompt_config(self):
         log_analysis = self.config.get("log_analysis", {})
-        system_prompt_first = log_analysis.get("system_prompt_first")
+        return {"system_prompt_first": log_analysis.get("system_prompt_first", ""),
+                "system_prompt_default": log_analysis.get("system_prompt_default", "")}
+
+    def get_alarm_mcp_config(self):
+        mcp = self.config.get("alarm_analysis", {}).get("mcp", {})
+        return {"mcp_grafana_url": mcp.get("mcp_grafana_url", ""), "mcp_mariadb_url": mcp.get("mcp_mariadb_url", ""),
+                "mcp_influxdb_url": mcp.get("mcp_influxdb_url", "")}
+
+    def get_alarm_system_prompt_config(self):
+        alarm_analysis = self.config.get("alarm_analysis", {})
+        system_prompt_first = alarm_analysis.get("system_prompt_first")
         if not system_prompt_first:
-            system_prompt_first = log_analysis.get("mcp", {}).get("system_prompt_first", "")
-        return {"system_prompt_first": system_prompt_first, "system_prompt_default": log_analysis.get("system_prompt_default", "")}
+            system_prompt_first = alarm_analysis.get("mcp", {}).get("system_prompt_first", "")
+        return {"system_prompt_first": system_prompt_first,
+                "system_prompt_default": alarm_analysis.get("system_prompt_default", "")}
